[AUD-1549] Move trending-underground to common
